+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactor tests to use vi.runAllTimers for consistent timer handling
 - Refactor effect that called simulateLoading to only trigger random error if in DEV_MODE
+- Remove unnecessary aria-live attributes from loading and error state components
 
 ---